fixing keyboard events
