+++ app/_components/_tdle/guessBoard.tsx
@@ -1,24 +1,11 @@
 'use client'
 
-import { PureComponent, useEffect } from "react";
 import CharRow from "./charRow"
 import { useTdleContext } from "./tdleContext"
 
 export default function GuessBoard() {
 
-    const { guesses, handleInput } = useTdleContext();
-
-    useEffect(() => {
-        const handleKeyDown = (event: KeyboardEvent) => {
-            handleInput(event.key.toLowerCase());
-        };
-
-        window.addEventListener('keyup', handleKeyDown);
-
-        return () => {
-            window.removeEventListener('keyup', handleKeyDown);
-        };
-    }, []);
+    const { guesses } = useTdleContext();
 
     return <div className="p-5 mx-auto">
         {guesses.map((guess, i) => {
